moved things around to resemble the sf way
